# 8732 rn r form generation is slow