--- a/server/service/src/rnr_form/tests/generate_rnr_form_lines.rs
+++ b/server/service/src/rnr_form/tests/generate_rnr_form_lines.rs
@@ -1,10 +1,16 @@
 #[cfg(test)]
 mod generate_rnr_form_lines {
+    use crate::rnr_form::generate_rnr_form_lines::{
+        generate_rnr_form_lines, get_adjusted_quantity_consumed, get_amc,
+        get_bulk_opening_balances, get_earliest_expiry, get_previous_monthly_consumption,
+        get_stock_out_duration, get_usage_map, UsageStats,
+    };
+    use crate::service_provider::ServiceProvider;
     use chrono::NaiveDate;
     use repository::mock::{
-        item_query_test1, mock_item_a, mock_master_list_program_b, mock_name_invad,
-        mock_period_2_a, mock_period_2_b, mock_period_2_c, mock_period_2_d, mock_program_b,
-        mock_rnr_form_a, MockData,
+        item_query_test1, item_query_test2, mock_item_a, mock_master_list_program_b,
+        mock_name_invad, mock_period_2_a, mock_period_2_b, mock_period_2_c, mock_period_2_d,
+        mock_program_b, mock_rnr_form_a, MockData,
     };
     use repository::mock::{mock_store_a, MockDataInserts};
     use repository::test_db::setup_all_with_data;
@@ -13,13 +19,7 @@ mod generate_rnr_form_lines {
         RnRFormFilter, RnRFormLineRow, RnRFormLowStock, RnRFormRow, StockLineRow,
         StorePreferenceRow, StorePreferenceRowRepository,
     };
-
-    use crate::rnr_form::generate_rnr_form_lines::{
-        generate_rnr_form_lines, get_adjusted_quantity_consumed, get_amc, get_earliest_expiry,
-        get_opening_balance, get_previous_monthly_consumption, get_stock_out_duration,
-        get_usage_map, UsageStats,
-    };
-    use crate::service_provider::ServiceProvider;
+    use std::collections::HashMap;
 
     #[actix_rt::test]
     async fn test_generate_rnr_form_lines() {
@@ -204,7 +204,6 @@ mod generate_rnr_form_lines {
                     // +5
                     invoice_line_inbound(),
                 ],
-                // Current stock on hand for item, 10 packs
                 stock_lines: vec![StockLineRow {
                     item_link_id: item_query_test1().id,
                     store_id: mock_store_a().id,
@@ -217,30 +216,28 @@ mod generate_rnr_form_lines {
         )
         .await;
 
-        // When previous row provided, use that value
-        let result = get_opening_balance(
-            &connection,
-            Some(&RnRFormLineRow {
+        let mut previous_lines = HashMap::new();
+        previous_lines.insert(
+            item_query_test2().id.clone(),
+            RnRFormLineRow {
                 final_balance: 7.0,
                 ..Default::default()
-            }),
-            &mock_store_a().id,
-            &item_query_test1().id,
-            NaiveDate::from_ymd_opt(2024, 1, 1).unwrap(),
-        )
-        .unwrap();
-        assert_eq!(result, 7.0);
+            },
+        );
+        let item_ids = vec![item_query_test1().id.clone(), item_query_test2().id.clone()];
 
-        // When no previous row, calculate as of starting date
-        let result = get_opening_balance(
+        let opening_balances = get_bulk_opening_balances(
             &connection,
-            None,
+            &previous_lines,
             &mock_store_a().id,
-            &item_query_test1().id,
+            &item_ids,
             NaiveDate::from_ymd_opt(2024, 1, 1).unwrap(),
         )
         .unwrap();
-        assert_eq!(result, 8.0);
+
+        assert_eq!(opening_balances.len(), 2);
+        assert_eq!(opening_balances.get(&item_query_test2().id).unwrap(), &7.0);
+        assert_eq!(opening_balances.get(&item_query_test1().id).unwrap(), &8.0);
     }
 
     #[actix_rt::test]