<!-- IMPORTANT!
  - Every PR must reference an issue; this helps to explain the intent of the PR
 -->

Fixes #8732

# 👩🏻‍💻 What does this PR do?
Massive performance improvements for R&R form generations!

### With backend tests:
**Original:** Form with 1000 items took: 6.055736166s
**Now:** Form with 1000 items took: 350.028167ms

### Datafile on tablet
**Now:** 4.5 seconds with 800 items on tablet
**Original:** Don’t know database keeps locking

Reduced loop (n+1):
- Store preferences was being called every time in the loop, have moved this out since we only need to call this once to get the store preferences...
- Getting opening balance in bulk outside of the loop
- Major improvement 
   - Stock out durations was using `get_stock_evolution_for_item` which was calculating also future projects and doing an extra database call we didn't need. Also moved out of the loop and getting in bulk
   - Also getting earliest expiry dates outside of loop 

Also added a test with 1k items and consumption/amc data which will fail if the generation takes longer than 2seconds! 

## 💌 Any notes for the reviewer?
We can do further performance improvements through sql calls and probably batch generation? Things to explore for the future (: This is already a huge improvement!

# 🧪 Testing

<!-- Explain the steps you'd take to test the changes of this PR manually -->

- [ ] Have a huge well used datafile with 1k + items
- [ ] Try generating R&R forms
- [ ] Should be fairly quick... maybe about 5 seconds? 👀 idk

# 📃 Documentation

- [ ] **Part of an epic**: documentation will be completed for the feature as a whole
- [ ] **No documentation required**: no user facing changes or a bug fix which isn't a change in behaviour
- [ ] **These areas should be updated or checked**: <!-- _(e.g.)_ New `issued` column in `Requisitions` indicates stock quantity already in shipments -->
  1.
  2.


# 📃 Reviewer Checklist

The PR Reviewer(s) should fill out this section before approving the PR

**Breaking Changes**
- [ ] No Breaking Changes in the Graphql API
- [ ] Technically some Breaking Changes but not expected to impact any integrations

**Issue Review**
- [ ] All requirements in original issue have been covered
- [ ] A follow up issue(s) have been created to cover additional requirements

**Tests Pass**
- [ ] Postgres
- [ ] SQLite
- [ ] Frontend